Change in functions _getenv exec and getline to print only execution of the commands
