+++ _getenv.c~
@@ -1,15 +1,15 @@
-#include "holberton.h"
 /**
  * _getenv - returns a pointer to the corresponding
  * value string of a environment variable.
  * @name: pointer to string-name of the require variable.
  * Return: a pointer to the string of the value or NULL.
  */
-char *_getenv(char *name)
+char *_getenv(const char *name)
 {
 	int i = 0, b = 0, c = 0;
 	char **test = environ;
 
+	printf("name:%s \n", name);
 	while (test[i])
 	{
 		b = 0;
@@ -24,10 +24,10 @@ char *_getenv(char *name)
 			}
 			b++;
 		}
-		if (_strlen(name) == c && test[i][b] == '=')
+		if (strlen(name) == c && test[i][b] == '=')
 		{
 			strtok(environ[i], "=");
-			return (strtok(NULL, "="));
+			return(strtok(NULL, "="));
 		}
 		i++;
 	}
