+++ exec.c
@@ -20,8 +20,8 @@ void fun_exec(char **commands)
 	store_paths = NULL;
 	pathstr = _getenv("PATH");
 	store_paths = _extractpath(&head, pathstr);
-	print_list(*store_paths);
-
+	/*print_list(*store_paths);*/
+	(void)store_paths;
 	/*create child proccess */
 	new_id = fork();
 	if (new_id == 0)
@@ -29,7 +29,7 @@ void fun_exec(char **commands)
 		if (commands[0][0] != '/')
 		{
 			alone = commands[0];
-			commands[0] = strcat(bin, alone);
+			commands[0] = _strcat(bin, alone);
 		}
 		/*execute process*/
 		child_id = getpid();
@@ -48,31 +48,3 @@ void fun_exec(char **commands)
 		wait(&child_id);
 	}
 }
-	char *_strcat(char *dest, char *src)
-	{
-		int destlen = 0;
-		int srclen = 0;
-		int i = 0;
-		char b;
-
-		while (*(dest + destlen) != '\0')
-		{
-			destlen++;
-		}
-		while (*(src + srclen) != '\0')
-		{
-			srclen++;
-		}
-		for (i = 0; i < srclen; i++)
-		{
-			if (*(src + i) != '\0')
-			{
-				b = destlen + i;
-				*(dest + b) = *(src + i);
-			}
-		}
-		b = destlen + srclen;
-		*(dest + b) = '\0';
-		printf("Dest;%s", dest);
-		return (dest);
-	}
